Adjust coop wg test params for bench / validation
