+++ test/gemm/device/mma_sync_multi_lds.hpp
@@ -234,7 +234,7 @@ namespace rocwmma
         }
         else
         {
-            GemmDriver::fill(fragsC, 0);
+            GemmDriver::fill(fragsC, static_cast<OutputT>(0));
         }
 
         ///
