+++ test/gemm/test/mma_sync_coop_wg_test_16x16_nn_2x2.cpp
@@ -41,7 +41,7 @@ namespace rocwmma
         // Types: ALL + double
         // Block Sizes: 16 x 16 x BlockK
         // Layouts: NN
-        using Types = typename Base::TestTypesIOC;
+        using Types = typename Base::TestTypes16x16;
         using BlockSizes
             = std::tuple<std::tuple<I<16>, I<16>, I<16>>, std::tuple<I<16>, I<16>, I<32>>>;
         using Layouts     = typename Base::TestLayoutsNN;
