+++ test/gemm/test/mma_sync_coop_wg_test_16x16_tt_8x4.cpp
@@ -26,7 +26,7 @@
 
 #include <type_traits>
 
-#include "detail/mma_sync_multi_lds.hpp"
+#include "detail/mma_sync_coop_wg.hpp"
 #include "gemm_config.hpp"
 #include "gemm_test.hpp"
 #include "kernel_generator.hpp"
@@ -38,22 +38,22 @@ namespace rocwmma
     {
         using Base = CommonTestParams;
 
-        // Types: Small sizes
+        // Types: Small types
         // Block Sizes: 16 x 16 x BlockK
-        // Layouts: NN
+        // Layouts: TT
         using Types       = typename Base::TestTypesSmall;
         using BlockSizes  = std::tuple<std::tuple<I<16>, I<16>, I<16>>>;
-        using Layouts     = typename Base::TestLayoutsNN;
+        using Layouts     = typename Base::TestLayoutsTT;
         using LayoutsLds  = typename Base::TestLdsLayoutTypes;
-        using MappingsLds = typename Base::TestMappingsLds;
-        using BlocksXY    = std::tuple<std::tuple<I<8>, I<8>>>;
+        using GemmConfigs = typename Base::TestGemmConfigsWgLevel;
+        using BlocksXY    = std::tuple<std::tuple<I<8>, I<4>>>;
         using KernelParams =
-            typename CombineLists<Types, BlockSizes, Layouts, LayoutsLds, MappingsLds, BlocksXY>::
+            typename CombineLists<Types, BlockSizes, Layouts, LayoutsLds, GemmConfigs, BlocksXY>::
                 Result;
 
         // Assemble the kernel generator
         // Kernel: MmaSyncMulti
-        using GeneratorImpl   = MmaSyncMultiLdsGenerator;
+        using GeneratorImpl   = MmaSyncCoopWgGenerator;
         using KernelGenerator = KernelGenerator<KernelParams, GeneratorImpl>;
 
         // Sanity check for kernel generator
@@ -69,18 +69,18 @@ namespace rocwmma
 } // namespace rocwmma
 
 // Test suite for unique parameterization
-class MmaSyncMultiLdsTest16x16NN8x8 : public rocwmma::GemmTest
+class MmaSyncCoopWgTest16x16TT8x8 : public rocwmma::GemmTest
 {
 };
 
-TEST_P(MmaSyncMultiLdsTest16x16NN8x8, RunKernel)
+TEST_P(MmaSyncCoopWgTest16x16TT8x8, RunKernel)
 {
     this->RunKernel();
 }
 
 INSTANTIATE_TEST_SUITE_P(
     GemmKernelTests,
-    MmaSyncMultiLdsTest16x16NN8x8,
+    MmaSyncCoopWgTest16x16TT8x8,
     ::testing::Combine(::testing::ValuesIn(rocwmma::TestParams::kernels()),
                        ::testing::ValuesIn(rocwmma::TestParams::threadBlocks()),
                        ::testing::ValuesIn(rocwmma::TestParams::problemSizes()),
