+++ test/gemm/test/mma_sync_coop_wg_test_32x32_nn_4x4.cpp
@@ -38,12 +38,11 @@ namespace rocwmma
     {
         using Base = CommonTestParams;
 
-        // Types: ALL + double
+        // Types: Small types
         // Block Sizes: 32 x 32 x BlockK
         // Layouts: NN
-        using Types = typename Base::TestTypes32x32;
-        using BlockSizes
-            = std::tuple<std::tuple<I<32>, I<32>, I<8>>, std::tuple<I<32>, I<32>, I<16>>>;
+        using Types       = typename Base::TestTypesSmall;
+        using BlockSizes  = std::tuple<std::tuple<I<32>, I<32>, I<8>>>;
         using Layouts     = typename Base::TestLayoutsNN;
         using LayoutsLds  = typename Base::TestLdsLayoutTypes;
         using GemmConfigs = typename Base::TestGemmConfigsWgLevel;
