+++ test/gemm/test/mma_sync_coop_wg_test_32x32_nt_1x1.cpp
@@ -38,7 +38,7 @@ namespace rocwmma
     {
         using Base = CommonTestParams;
 
-        // Types: ALL + double
+        // Types: ALL
         // Block Sizes: 32 x 32 x BlockK
         // Layouts: NT
         using Types       = typename Base::TestTypes32x32;
